Refactor tests base class
